requirements.txt
Django>=2.2.13,<3
djangorestframework
psycopg2

flake8
isort
ipdb